requirements.txt
amqp==5.0.6
atpublic==1.0
billiard==3.6.4.0
celery==5.2.1
certifi==2018.11.29
coverage==4.5.2
chardet==3.0.4
Django==3.2.18
django-celery-beat==1.5.0
django-ckeditor==5.4.0
django-excel==0.0.10
django-js-asset==1.1.0
#django-multi-captcha-admin==1.0.0
django-recaptcha2==1.3.1
django-timezone-field==3.0
django-jet==1.0.8
et-xmlfile==1.0.1
Faker==0.8.8
flufl.bounce==3.0
idna==2.8
jdcal==1.4
kombu==5.2.2
lml==0.0.7
mixer==6.0.1
openpyxl==2.5.12
Pillow==9.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyexcel==0.5.10
pyexcel-io==0.5.11
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
python-crontab==2.3.6
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
six==1.12.0
text-unidecode==1.2
texttable==1.5.0
urllib3==1.24.1
vine==5.0.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.0.0
python-dotenv==1.0.0
dj-database-url==2.1.0
whitenoise==6.6.0
psycopg2-binary==2.9.10
gunicorn==21.2.0
